API/requirements.txt
fastapi
uvicorn[standard]
aiofiles
jinja2
numpy
pandas
tensorflow-cpu
kaleido
